Move Inferencing out of the Typer cake
